pyproject.toml
[project]
name = "mcp-shield"
version = "0.1.0"
description = "Secure MCP runtime with policy enforcement and sandboxing"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "typer>=0.12",
    "mcp>=1.0",
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff>=0.4", "black>=24.0"]
